fix: rename services in code
