+++ web/src/lib/ui/commands/saveLayoutCommand.ts
@@ -1,6 +1,6 @@
 import { produce } from "immer";
-import makeLayoutsStatic from "~/client/services/makeLayoutsStaticService";
-import { transformRGLToAWL } from "~/client/services/transformLayoutsService";
+import makeLayoutsStatic from "~/client/services/makeLayoutsStatic.service";
+import { transformRGLToAWL } from "~/client/services/transformLayouts.service";
 import { type RGLayout } from "~/lib/types/widget";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
 import { useBoundStore } from "../state";
