+++ web/src/lib/ui/hooks.ts
@@ -1,8 +1,8 @@
 import { useIsClient, useWindowSize } from "@uidotdev/usehooks";
 import { useAtom } from "jotai";
 import { useContext, useEffect, useState } from "react";
-import filterWidgetLayoutByLayout from "~/client/services/filterWidgetLayoutByLayoutService";
-import getHidingClasses from "~/client/services/getHidingClassesService";
+import filterWidgetLayoutByLayout from "~/client/services/filterWidgetLayoutByLayout.service";
+import getHidingClasses from "~/client/services/getHidingClasses.service";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
 import { api } from "../api/api";
 import { BREAKPOINTS } from "../basic/const";
