+++ web/src/server/domain/layout/repo/layoutUpstashRepository.ts
@@ -3,7 +3,7 @@ import { env } from "~/env.mjs";
 import { UPSTASH_LAYOUT_KEY } from "~/lib/basic/const";
 import { codes } from "~/lib/error/codes";
 import AppError from "~/lib/error/error";
-import { parseAdjustedWidgetLayout } from "../../config/services/parseWidgetConfigService";
+import { parseAdjustedWidgetLayout } from "../../config/services/parseWidgetConfig.service";
 import type { AdjustedWidgetLayout } from "../adjustedWidgetLayout";
 import type { LayoutRepository } from "./layoutRepository";
